reorganizing files order and check routes compatibility with React FE
