+++ package.json
@@ -21,6 +21,7 @@
   "dependencies": {
     "algoliasearch": "^4.8.0",
     "bcrypt": "^5.0.0",
+    "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
