+++ src/app.js
@@ -1,29 +1,16 @@
 const express = require('express')
-const app = express()
 const expressLayouts = require('express-ejs-layouts')
-
 const config = require('dotenv'); // .env file
 
 // single database connection shared to all routes
 // other routes can access using getDb
-const initDb = require("./db/db.js").initDb;
-const getDb = require("./db/db.js").getDb;
-// const { registerRoutes } = require('./routes');
-
-const indexRouter=require('./routes/index.js')
-const projectsRouter=require('./routes/projects.js')
+const initDb = require("./db/db.js").initDb
+const getDb = require("./db/db.js").getDb
+const bodyParser = require("body-parser")
 
 if (process.env.NODE_ENV !== 'production') {
     require('dotenv').config();
 }
-
-
-app.set('view engine','ejs')
-app.set('views', __dirname + '/views')
-app.set('layout','layouts/layout')
-app.use(expressLayouts)
-app.use(express.static('public'))
-
 
 const port = 3001;
 
@@ -36,6 +23,21 @@ initDb(function (err) {
     })}
 );
 
+
+const indexRouter = require('./routes/index.js')
+const projectsRouter = require('./routes/projects.js')
+
+const app = express()
+
+
+app.set('view engine','ejs')
+app.set('views', __dirname + '/views')
+app.set('layout','layouts/layout')
+app.use(expressLayouts)
+app.use(express.static('public'))
+app.use(bodyParser.urlencoded({ extended: true }));
+
+
 app.use('/index', indexRouter)
 app.get('/', (req, res) => {
   res.send('NEXUS UW App Backend 1')
@@ -44,4 +46,5 @@ app.use('/', indexRouter);
 app.use('/projects', projectsRouter);
 
 
-app.listen(process.env.PORT || 3000), () => console.log(`Server running on port ${process.env.PORT}`);
+app.listen(process.env.PORT || 3000), ()
+    => console.log(`Server running on port ${process.env.PORT}`);
